#677 fixing naming and moved to container tests
